requirements.txt
asgiref==3.8.1
Django==5.1
django-filter==24.3
django-taggit==6.0.0
djangorestframework==3.15.2
pillow==10.4.0
sqlparse==0.5.1
